LPS-59429 rename
