+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/lifecycle/CheckJournalContentSearchPortalInstanceLifecycleListener.java
@@ -12,23 +12,19 @@
  * details.
  */
 
-package com.liferay.journal.events;
+package com.liferay.journal.lifecycle;
 
-import com.liferay.dynamic.data.mapping.util.DefaultDDMStructureHelper;
-import com.liferay.journal.model.JournalArticle;
-import com.liferay.journal.service.JournalArticleLocalService;
+import com.liferay.journal.configuration.JournalServiceConfigurationValues;
+import com.liferay.journal.service.JournalContentSearchLocalService;
 import com.liferay.portal.kernel.events.ActionException;
 import com.liferay.portal.kernel.events.SimpleAction;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.model.Company;
-import com.liferay.portal.model.Group;
 import com.liferay.portal.security.auth.CompanyThreadLocal;
 import com.liferay.portal.service.CompanyLocalService;
-import com.liferay.portal.service.GroupLocalService;
-import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.UserLocalService;
-import com.liferay.portal.util.PortalUtil;
 
 import java.util.List;
 
@@ -40,7 +36,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eudaldo Alonso
  */
 @Component(immediate = true)
-public class AddDefaultJournalStructuresAction extends SimpleAction {
+public class CheckJournalContentSearchPortalInstanceLifecycleListener extends SimpleAction {
 
 	@Override
 	public void run(String[] ids) throws ActionException {
@@ -71,28 +67,15 @@ public class AddDefaultJournalStructuresAction extends SimpleAction {
 	}
 
 	protected void doRun(long companyId) throws Exception {
-		ServiceContext serviceContext = new ServiceContext();
-
-		serviceContext.setAddGuestPermissions(true);
-		serviceContext.setAddGroupPermissions(true);
-
-		Group group = _groupLocalService.getCompanyGroup(companyId);
-
-		serviceContext.setScopeGroupId(group.getGroupId());
-
-		long defaultUserId = _userLocalService.getDefaultUserId(companyId);
-
-		serviceContext.setUserId(defaultUserId);
-
-		Class<?> clazz = getClass();
-
-		_defaultDDMStructureHelper.addDDMStructures(
-			defaultUserId, group.getGroupId(),
-			PortalUtil.getClassNameId(JournalArticle.class),
-			clazz.getClassLoader(),
-			"com/liferay/journal/upgrade/v1_0_0/dependencies" +
-				"/basic-web-content-structure.xml",
-			serviceContext);
+		if (JournalServiceConfigurationValues.SYNC_CONTENT_SEARCH_ON_STARTUP) {
+			try {
+				_journalContentSearchLocalService.checkContentSearches(
+					companyId);
+			}
+			catch (Exception e) {
+				_log.error(e, e);
+			}
+		}
 	}
 
 	@Reference(unbind = "-")
@@ -103,20 +86,10 @@ public class AddDefaultJournalStructuresAction extends SimpleAction {
 	}
 
 	@Reference(unbind = "-")
-	protected void setDefaultDDMStructureHelper(
-		DefaultDDMStructureHelper defaultDDMStructureHelper) {
+	protected void setJournalContentSearchLocalService(
+		JournalContentSearchLocalService journalContentSearchLocalService) {
 
-		_defaultDDMStructureHelper = defaultDDMStructureHelper;
-	}
-
-	@Reference(unbind = "-")
-	protected void setGroupLocalService(GroupLocalService groupLocalService) {
-		_groupLocalService = groupLocalService;
-	}
-
-	@Reference(unbind = "-")
-	protected void setJournalArticleLocalService(
-		JournalArticleLocalService journalArticleLocalService) {
+		_journalContentSearchLocalService = journalContentSearchLocalService;
 	}
 
 	@Reference(target = ModuleServiceLifecycle.PORTAL_INITIALIZED, unbind = "-")
@@ -124,14 +97,10 @@ public class AddDefaultJournalStructuresAction extends SimpleAction {
 		ModuleServiceLifecycle moduleServiceLifecycle) {
 	}
 
-	@Reference(unbind = "-")
-	protected void setUserLocalService(UserLocalService userLocalService) {
-		_userLocalService = userLocalService;
-	}
+	private static final Log _log = LogFactoryUtil.getLog(
+		CheckJournalContentSearchPortalInstanceLifecycleListener.class);
 
 	private CompanyLocalService _companyLocalService;
-	private DefaultDDMStructureHelper _defaultDDMStructureHelper;
-	private GroupLocalService _groupLocalService;
-	private UserLocalService _userLocalService;
+	private JournalContentSearchLocalService _journalContentSearchLocalService;
 
 }
